pyproject.toml
[project]
name = "cellprofiler-dspy-review"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "dspy-ai>=3.2.0",
    "httpx>=0.28.1",
    "pydantic>=2.13.3",
]
